pyproject.toml
[tool.poetry]
name = "learning-sam-vit"
version = "0.1.0"
description = ""
authors = ["gomesuit <9404411+gomesuit@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "learning_sam_vit"}]

[tool.poetry.dependencies]
python = "^3.10"
transformers = "^4.35.2"
pillow = "^10.1.0"
torch = "^2.1.1"
torchvision = "^0.16.1"
matplotlib = "^3.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
